ingots_tools/cuttleagent/pyproject.toml
[project]
name = "cuttleagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "langchain>=1.2.10",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.2",
    "typer>=0.24.1",
    "kexploit_agent",
]

[tool.uv]
package = true

[tool.uv.sources]
kexploit_agent = { workspace = true }

[project.scripts]
cuttleagent = "cuttleagent.main:main"
